Adjust FB login flow
