+++ web/portal/src/config.js
@@ -2,7 +2,7 @@
 
 // setup server url
 const Config = {
-    baseURL: 'https://r2c5wmub95.execute-api.ap-northeast-1.amazonaws.com/devjim/'
+    baseURL: 'https://r2c5wmub95.execute-api.ap-northeast-1.amazonaws.com/devjim'
 };
 
 export default Config;
